Updated to check for user in required routes.

+++ server/src/Services/Post/index.ts
@@ -3,6 +3,7 @@ import { InjectRepository } from 'typeorm-typedi-extensions';
 import PostRepository from '../../Repository/Posts';
 import { Post } from '../../Models/Post';
 import { ObjectLiteral } from 'typeorm';
+import { User } from '../../../types';
 
 @Service()
 export default class PostService {
@@ -19,9 +20,15 @@ export default class PostService {
 
   public updatePost = async (
     id: number,
-    postData: any
+    postData: any,
+    user: User
   ): Promise<any | Error> => {
-    // Todo: get user and check if same as author.
+    const currentPost = await this.postRepository.getById(id);
+
+    if (currentPost.id !== user?.id) {
+      return Error(`You are not authorised to perform this action.`);
+    }
+
     const { affected } = await this.postRepository.update(id, postData);
     if (!affected) {
       return Error(`Could not update post with specified id:${id}.`);
@@ -38,8 +45,16 @@ export default class PostService {
     return post;
   };
 
-  public deletePost = async (id: number): Promise<ObjectLiteral | Error> => {
-    // Todo: get user and check if same as author.
+  public deletePost = async (
+    id: number,
+    user: User
+  ): Promise<ObjectLiteral | Error> => {
+    const currentPost = await this.postRepository.getById(id);
+
+    if (currentPost.id !== user?.id) {
+      return Error(`You are not authorised to perform this action.`);
+    }
+
     const post = await this.postRepository.delete(id);
     if (!post) {
       return Error(`Post with specified id:${id} not found.`);
